Struct docgen::DocgenOptions

pub struct DocgenOptions {

    pub section_level_start: usize,
    pub include_private_fun: bool,
    pub include_specs: bool,
    pub specs_inlined: bool,
    pub include_impl: bool,
    pub toc_depth: usize,
    pub collapsed_sections: bool,
    pub output_directory: String,
    pub doc_path: Vec<String>,
    pub root_doc_templates: Vec<String>,
    pub references_file: Option<String>,
    pub include_dep_diagrams: bool,
    pub include_call_diagrams: bool,
    pub compile_relative_to_output_dir: bool,

}

Options passed into the documentation generator.

Fields
section_level_start: usize

The level where we start sectioning. Often markdown sections are rendered with unnecessary large section fonts, setting this value high reduces the size.

include_private_fun: bool

Whether to include private functions in the generated docs.

include_specs: bool

Whether to include specifications in the generated docs.

specs_inlined: bool

Whether to put specifications in the same section as a declaration or put them all into an independent section.

include_impl: bool

Whether to include Move implementations.

toc_depth: usize

Max depth to which sections are displayed in table-of-contents.

collapsed_sections: bool

Whether to use collapsed sections (
) for impl and specs

output_directory: String

In which directory to store output.

doc_path: Vec<String>

In which directories to look for references.

root_doc_templates: Vec<String>

A list of paths to files containing templates for root documents for the generated documentation.

A root document is a markdown file which contains placeholders for generated documentation content. It is also processed following the same rules than documentation comments in Move, including creation of cross-references and Move code highlighting.

A placeholder is a single line starting with a markdown quotation marker of the following form:

> {{move-include NAME_OF_MODULE_OR_SCRIPT}}
> {{move-toc}}
> {{move-index}}

These lines will be replaced by the generated content of the module or script, or a table of contents, respectively.

For a module or script which is included in the root document, no separate file is generated. References between the included and the standalone module/script content work transparently.

references_file: Option<String>

An optional file containing reference definitions. The content of this file will be added to each generated markdown doc.

include_dep_diagrams: bool

Whether to include dependency diagrams in the generated docs.

include_call_diagrams: bool

Whether to include call diagrams in the generated docs.

compile_relative_to_output_dir: bool

If this is being compiled relative to a different place where it will be stored (output directory).